requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
anyio==4.8.0
    # via httpx
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
h11==0.14.0
    # via httpcore
httpcore==1.0.7
    # via httpx
httpx==0.28.1
    # via python-telegram-bot
idna==3.10
    # via
    #   anyio
    #   httpx
python-telegram-bot==21.11.1
    # via woitelegram-bot (pyproject.toml)
sniffio==1.3.1
    # via anyio
tornado==6.4.2
    # via python-telegram-bot
typing-extensions==4.12.2
    # via anyio
